+++ guillotina/test_package.py
@@ -1,5 +1,6 @@
 # this is for testing.py, do not import into other modules
 from guillotina import configure
+from guillotina import fields
 from guillotina import schema
 from guillotina.async_util import IAsyncUtility
 from guillotina.behaviors.instance import AnnotationBehavior
@@ -9,6 +10,7 @@ from guillotina.content import Item
 from guillotina.content import Resource
 from guillotina.directives import index_field
 from guillotina.directives import metadata
+from guillotina.directives import read_permission
 from guillotina.directives import write_permission
 from guillotina.exceptions import NoIndexField
 from guillotina.fields import CloudFileField
@@ -137,6 +139,16 @@ class IMarkerBehavior(Interface):
 class ITestBehavior(Interface):
     foobar = schema.TextLine(required=False)
     foobar_context = schema.TextLine(required=False, default='default-foobar')
+
+    bucket_dict = fields.BucketDictField(
+        bucket_len=10, required=False,
+        default=None,
+        key_type=schema.Text(),
+        value_type=schema.Text()
+    )
+
+    read_permission(no_read_field='example.MyPermission')
+    no_read_field = schema.TextLine(required=False, default='')
 
 
 @configure.behavior(